cloudfunction_scripts/requirements.txt
# Function dependencies, for example:
# package>=version
gcsfs==2022.11.0
nbformat==5.7.0
nbconvert==7.2.6
ipykernel==6.17.1

kfp==1.8.16
google-cloud-aiplatform==1.19.0
google-cloud-pipeline-components==1.0.29
datetime